added purse field to player table

+++ src/main/resources/sqls/patches/v3.0/updatePlayer.sql
@@ -2,4 +2,4 @@
  * Author:  rellu
  * Created: 21.01.2023
  */
-ALTER TABLE `rellu_essentials`.`player` ADD purse BIGINT(21) NOT NULL DEFAULT '0'; 
+ALTER TABLE `rellu_essentials`.`player` ADD purse FLOAT(21) NOT NULL DEFAULT '0';
